dev-requirements.txt
python-dotenv
psycopg2-binary
pre-commit
pytest
